+++ cli/docs/changelog.md
@@ -1,4 +1,10 @@
 # TBD
+### Features
+* `enclave` commands also show enclave state
+* Standardized table-printing logic into a `TablePrinter` object 
+* Added a `KeyValuePrinter` for pretty-printing key-value pairs
+* `enclave inspect` also prints the enclave ID & state
+
 ### Fixes
 * `module exec` will attempt to update the module & API container images before running